fix: inputTree as a String and not Object
BREAKING CHANGE

For large trees, using inputTree props instead of newick props highly decreased performance.
The explanation is that the inputTree object is transformed during by the d3 functions and this
is not compatible with the "not mutable" constraint of the props...
I tested cloneDeepp to clone the input object but it took too much time.
Same thing for JSON.Sparse - JSON.Stringify.

In the app using vue-phylogram, you thus have to move inputTree objects to JSON strings.